# StoreGateway: Parallelize Object Storage Requests While Building Index Headers