--- a/pkg/storegateway/indexheader/header_test.go
+++ b/pkg/storegateway/indexheader/header_test.go
@@ -11,6 +11,7 @@ import (
 	"fmt"
 	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/go-kit/log"
 	"github.com/grafana/dskit/gate"
@@ -25,6 +26,7 @@ import (
 	"github.com/thanos-io/objstore"
 	"github.com/thanos-io/objstore/providers/filesystem"
 
+	"github.com/grafana/mimir/pkg/storage/bucket"
 	"github.com/grafana/mimir/pkg/storage/tsdb/block"
 	"github.com/grafana/mimir/pkg/util/test"
 )
@@ -416,6 +418,35 @@ func BenchmarkBinaryWrite(t *testing.B) {
 	t.ResetTimer()
 	for i := 0; i < t.N; i++ {
 		require.NoError(t, WriteBinary(ctx, bkt, m.ULID, fn))
+	}
+}
+
+func BenchmarkBinaryWrite_DelayedBucket(t *testing.B) {
+
+	tests := []struct {
+		delay time.Duration
+	}{
+		{10 * time.Millisecond},
+		{20 * time.Millisecond},
+		{50 * time.Millisecond},
+	}
+
+	for _, test := range tests {
+		t.Run(fmt.Sprintf("max_delay_%d_ms", test.delay.Milliseconds()), func(t *testing.B) {
+			ctx := context.Background()
+			tmpDir := t.TempDir()
+			bkt, err := filesystem.NewBucket(filepath.Join(tmpDir, "bkt"))
+			require.NoError(t, err)
+			delaybkt := bucket.NewDelayedBucketClient(bkt, 0, test.delay)
+			defer func() { require.NoError(t, delaybkt.Close()) }()
+
+			m := prepareIndexV2Block(t, tmpDir, delaybkt)
+			fn := filepath.Join(tmpDir, m.ULID.String(), block.IndexHeaderFilename)
+			t.ResetTimer()
+			for i := 0; i < t.N; i++ {
+				require.NoError(t, WriteBinary(ctx, delaybkt, m.ULID, fn))
+			}
+		})
 	}
 }