#### What this PR does

- Improves the performance of `WriteBinary` by refactoring two sequential calls to object storage to run in parallel. The Store Gateway calls `syncBlocks` on startup and periodically while running (see: `blocks-storage.bucket-store.sync-interval`)  to refresh the bucket index and look for new blocks shipped by ingesters or blocks deleted by retention/ compaction. Part of this sync involves calling `WriteBinary` which updates a local index-header file using pieces of the index in object storage. The requests for these pieces of the index file can be sent in parallel if we write them to the local index-header in the correct order.

- Adds a new benchmark, `WriteBinary_DelayedBucket`, which uses a mock bucket with a built-in delay to simulate calls to remote providers. 

```text
pkg: github.com/grafana/mimir/pkg/storegateway/indexheader
                                            │   old.txt    │               new.txt               │
                                            │    sec/op    │   sec/op     vs base                │
WriteBinary-8                                 9.156m ±  1%   9.994m ± 5%   +9.15% (p=0.000 n=10)
WriteBinary_DelayedBucket/max_delay_10_ms-8   66.72m ±  4%   58.33m ± 4%  -12.57% (p=0.000 n=10)
WriteBinary_DelayedBucket/max_delay_20_ms-8   120.8m ±  8%   105.4m ± 4%  -12.72% (p=0.000 n=10)
WriteBinary_DelayedBucket/max_delay_50_ms-8   285.8m ± 11%   224.3m ± 9%  -21.53% (p=0.000 n=10)
```

#### Which issue(s) this PR fixes or relates to

#8166

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.