@dmwilson-grafana 2025-02-20
> I'm not 100% convinced this will give us meaningful speedups for larger index headers. The GetRange readers we get buffer only about 64K of data. For for larger index headers where delay actually matters (where downloading the index header is in the seconds/minutes) I think this will perform effectively the same.

You're right. With this change we only get an improvement that's roughly the round-trip time to object storage. The original benchmarks used a 6MB index, when I tested w. a 2.2GB block (see below) we see about the same improvement in ms and almost no improvement in %'age terms. 


```
                                            │ regular.txt  │          parallel.txt          │
                                            │    sec/op    │    sec/op     vs base          │
WriteBinary_DelayedBucket/max_delay_20_ms-8   483.8m ± 53%   470.7m ± 37%  ~ (p=0.853 n=10)
```